tests/wpt/tests/feature-policy/reporting/payment-reporting.https.html

<!DOCTYPE html>
<html>
  <head>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
    <script src='../resources/feature-policy-report-json.js'></script>
  </head>
  <body>
    <script>
var t = async_test("PaymentRequest Report Format");

var check_report_format = (reports, observer) => {
  let report = reports[0];
  assert_equals(report.type, "permissions-policy-violation");
  assert_equals(report.url, document.location.href);
  assert_equals(report.body.featureId, "payment");
  assert_equals(report.body.sourceFile, document.location.href);
  assert_equals(typeof report.body.lineNumber, "number");
  assert_equals(typeof report.body.columnNumber, "number");
  assert_equals(report.body.disposition, "enforce");
  check_report_json(report);
};

new ReportingObserver(t.step_func_done(check_report_format),
                      {types: ['permissions-policy-violation']}).observe();

t.step_func(() => {
  assert_throws_dom('SecurityError',
  () => new PaymentRequest(
    [{ supportedMethods: 'https://example.com/pay' }],
    { total: { label: 'Total', amount: { currency: 'USD', value: 0 }}},
    {}).show(),
  "PaymentRequest API should not be allowed in this document.");
})();
    </script>
  </body>
</html>
